[ADMINAPI-767] Fully reset EdOrgId's based on new list (#63)
* Fully reset EdOrgId's based on new list

* Run these workflows for hotfix PRs

+++ Application/EdFi.Ods.AdminApi/Features/Applications/EditApplication.cs
@@ -80,7 +80,7 @@ public class EditApplication : IFeature
         public IEnumerable<int>? ProfileIds { get; set; }
 
         [SwaggerSchema(Description = FeatureConstants.EducationOrganizationIdsDescription, Nullable = false)]
-        public IEnumerable<int>? EducationOrganizationIds { get; set; }
+        public IEnumerable<long>? EducationOrganizationIds { get; set; }
 
         [SwaggerSchema(Description = FeatureConstants.OdsInstanceIdDescription, Nullable = false)]
         public int OdsInstanceId { get; set; }
